+++ tests/test_commands/conftest.py
@@ -76,3 +76,18 @@ def base_timeseries_exog_path():
     tmp.flush()
     yield Path(tmp.name)
     tmp.close()
+
+
+@pytest.fixture
+def base_forecast_omegaconf_path():
+    tmp = NamedTemporaryFile("w")
+    tmp.write(
+        """
+        prediction_interval: true
+        quantiles: [0.025, 0.975]
+        n_folds: 3
+        """
+    )
+    tmp.flush()
+    yield Path(tmp.name)
+    tmp.close()